Adding and viewing database functional. Need to get some manner of onUpdate method as changes only show onCreate i.o.w have to close app and open it again in order to show changes.
Going to have to clean up a lot of code in the Home class which was used for demonstration as incompattible with database handler. Will do this one figured how to link in the existing edit and save buttons.

Suggest capturing return key input on the editTExt for details, as the save button is obscured by the textbox. Else move the save button elsewhere.

The list view still needs reformatting, displaying all relevant info, but in older format which would hide half of it.